requirements.txt
discord.py
python-dotenv
boto3
requests
